replace badges and random-projects page with tutorials page
The badges page was broken so let's just remove it all together. The
random projects page was a bit empty but people might still find it
through links so I'll just remove it from the header.

Instead, I'll add a new tutorials page. I hope to put more tutorials
here in the future so it's not such a graveyard as the random projects
page.