remove relativize_url
This is initially added for the IPFS mirror, which requires are urls to
be relative, as the website is prefixed with /ipfs/[hash]/. Since we
don't mirror anymore this is unnecessary.

It has now been removed as it made it impossible to add a custom 404
page. Nginx will render the /404.html page for all not-found resources.
The urls would be calculated relative to /404.html, instead of the
actual path the user visits, breaking all links.

It also removes a redirect when visited trailing-slash pages, as
relativize_url will remove all trailing-slashes.